Remove added temporary DOCTYPE
I'm using this library in my custom templates library. I read files with html pieces that do not need a concrete structure and that are going to be inserted inside another html text. Therefore, I need to maintain the original structure and not add extra elements. 
Therefore, if a tag or temporary element is added, the ideal is to erase it when it is no longer necessary.